针对一个图像块的预测变换量化逆变换逆量化已完成，但将上一个块的重建值传给下一个块时比较麻烦，由于考虑不能实现并行，边界扩展很繁琐实际需要扩展成9*9的块，准备更改方案，以大图为单位进行处理。一幅图存储重建值，一幅图存储原图

+++ intra_lss/DCTtransfer.cpp
@@ -1,6 +1,6 @@
 
 
-void DCTCore4x4(int16_t block[4][4], int16_t tblock[4][4])
+void DCTCore4x4(uint8_t block[4][4], uint8_t tblock[4][4])
 {
 	int16_t i, ii;
 	int16_t tmp[16];
@@ -50,7 +50,7 @@ void DCTCore4x4(int16_t block[4][4], int16_t tblock[4][4])
 	}
 }
 
-void IDCTCore4x4(int16_t tblock[4][4], int16_t block[4][4])
+void IDCTCore4x4(uint8_t tblock[4][4], uint8_t block[4][4])
 {
 	int16_t i, ii;
 	int16_t tmp[16];
